Update AutomatePackagingAndPatching.ps1 to include example of using PSOpenAI for certain parameters
Swapped commented out code for OpenAI API integration example. This is because the previously used PowerShellAI module was retired in June, 2024. The example now uses PSOpenAI. Read the comments if you would like to try the integration out.

Also updated notes at header of the script to reflect the change.